+++ projects/fhi-angular-components/src/lib/fhi-angular-components.module.ts
@@ -3,14 +3,14 @@ import { NgModule } from '@angular/core';
 import { FhiMultiselectModule } from './fhi-multiselect/fhi-multiselect.module';
 import { FhiAutosuggestModule } from './fhi-autosuggest/fhi-autosuggest.module';
 import { FhiTreeViewNavigationModule } from './fhi-tree-view-navigation/fhi-tree-view-navigation.module';
-import { FhiTreeViewCheckboxModule } from './fhi-tree-view-checkbox/fhi-tree-view-checkbox.module';
+import { FhiTreeViewSelectionModule } from './fhi-tree-view-selection/fhi-tree-view-selection.module';
 
 
 const FHI_ANGULAR_MODULES = [
   FhiAutosuggestModule,
   FhiMultiselectModule,
   FhiTreeViewNavigationModule,
-  FhiTreeViewCheckboxModule
+  FhiTreeViewSelectionModule
 ];
 
 @NgModule({