+++ projects/fhi-angular-components/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 > May 25, 2023
 
-* :tada: **Enhancement** Add support for single selection (radio) `FhiTreeViewCheckboxComponent`
+* :gift: **New** `FhiTreeViewSelectionComponent` is replacing `FhiTreeViewCheckboxComponent` :boom:.
+  * `FhiTreeViewSelectionComponent` supports single selection (radio).
 
 # 0.5.0